Second try of refactoring /resources
