+++ src/errors.rs
@@ -1,0 +1,6 @@
+error_chain! {
+    extern_links {
+        ImageError(::image::ImageError),
+        Io(::std::io::Error),
+    }
+}
